+++ backend/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "backend"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+repository.workspace = true
+readme.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+shadocal-lib = { path = "../shadocal-lib" }
+anyhow.workspace = true
+tokio = { workspace = true, features = ["full"] }
+
+actix-cors = "0.7.0"
+actix-web = "4.9.0"
