+++ backend/src/server.rs
@@ -1,14 +1,9 @@
 use actix_cors::Cors;
 use actix_web::{get, http, web, App, HttpResponse, HttpServer, Responder};
 
-pub mod state;
-mod tana;
-mod types;
+use shadocal_lib::State;
 
-pub use state::State;
-pub use types::*;
-
-use super::*;
+use super::tana;
 
 pub struct Server {
     ip: String,
@@ -29,8 +24,8 @@ impl Server {
     pub fn new(ip: String, port: u16) -> Self {
         Self { ip, port }
     }
-    pub async fn run(self, config: Option<Config>) -> anyhow::Result<()> {
-        let state = web::Data::new(State::new(config).await?);
+    pub async fn run(self, state: State) -> anyhow::Result<()> {
+        let state = web::Data::new(state);
 
         println!("🚀 Server started successfully");
         HttpServer::new(move || {
@@ -51,13 +46,4 @@ impl Server {
         .await?;
         Ok(())
     }
-}
-
-pub fn ip_port() -> (String, u16) {
-    let ip = std::env::var("SHADOCAL_IP");
-    let port = std::env::var("SHADOCAL_PORT").map(|v| {
-        v.parse::<u16>()
-            .expect("Invalid environment variable: SHADOCAL_PORT. Must be a valid port number.")
-    });
-    (ip.unwrap_or("127.0.0.1".to_string()), port.unwrap_or(7117))
 }
