+++ shadocal-lib/Cargo.toml
@@ -16,13 +16,7 @@ tokio.workspace = true
 
 chrono = { version = "0.4", features = ["serde"] }
 directories = "5.0"
-dotenv = "0.15.0"
-
-actix-cors = "0.7"
-actix-web = "4.9"
-url = "^2"
 
 gcal_rs = "0.1.5"
 serde = { version = "1.0.209", features = ["derive"] }
 async-trait = "0.1.82"
-toml = "0.8.19"
